+++ CHANGELOG.md
@@ -16,6 +16,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 - update jeap-messaging from 11.13.0 to 11.14.0
 - added ErrorHandlingTargetFilter to filter out messages not intended for the consuming service (Header: jeap_eh_target_service)
+- use ErrorHandlingTargetFilter for sequential inbox
 
 ## [14.13.0] - 2026-01-23
 ### Changed